requirements.txt
flask
requests
WTForms
flask-sqlalchemy
flask-login
flask-mail